feature: mapping number to bigint

+++ src/main/java/ch/admin/bar/siard2/jdbc/OracleMetaColumns.java
@@ -126,7 +126,7 @@ public class OracleMetaColumns
                             pt = PreType.BIGINT;
                     }
                 } else {
-                    pt = PreType.FLOAT;
+                    pt = PreType.BIGINT;
                     iDecimals = PredefinedType.iUNDEFINED;
                     iColumnSize = PredefinedType.iUNDEFINED;
                 }
